+++ tests/web/test_web_home.py
@@ -415,35 +415,6 @@ def test_info_send_when_not_filled(web_base_url, selenium):
 @test_case()
 @web
 @expected_failure
-def test_message_sent_when_filled(web_base_url, selenium):
-    """Test if message is sent when all info is filled."""
-    # GIVEN: On the OpenStax homepage
-    # WHEN: Go to bottom of the page and
-    # click on contact us in the footer with dark background
-    # AND: Fill out name, email and message
-    # AND: Click the orange send button
-    # THEN: The message is sent successfully.
-
-
-@test_case()
-@web
-@expected_failure
-def test_message_when__invalid(web_base_url, selenium):
-    """Test if message is sent when email is not valid."""
-    # GIVEN: On the OpenStax homepage
-    # WHEN: Go to bottom of the page and
-    # click on contact us in the footer with dark background
-    # AND: Fill out name, and message
-    # AND: For the email, put abcdedfg
-    # THEN: Message should not be sent and red string
-    # "Please include an '@' in the email address.
-    # 'abcdedfg is missing an '@'."
-    # should show up below the email input box.
-
-
-@test_case()
-@web
-@expected_failure
 def test_support_center(web_base_url, selenium):
     """Test if support center link works properly."""
     # GIVEN: On the OpenStax homepage
@@ -465,48 +436,6 @@ def test_box(web_base_url, selenium):
     # WHEN: Go to bottom of the page and
     # click on contact us in the footer with dark background
     # THEN: A box with green background with string "Contact Us" is displayed
-
-
-@test_case()
-@web
-@expected_failure
-def test_openstax_mailing_address(web_base_url, selenium):
-    """Test if openstax mailing address is properly displayed."""
-    # GIVEN: On the OpenStax homepage
-    # WHEN: Go to bottom of the page and
-    # click on contact us in the footer with dark background
-    # THEN: On the right column of the contact us page, mailing address of
-    # OpenStax is visible
-
-
-@test_case()
-@web
-@expected_failure
-def test_view_options_on_donating(web_base_url, selenium):
-    """Test if all donate options are availble."""
-    # GIVEN: On the OpenStax homepage
-    # WHEN: From the give page
-    # AND: Scroll down to the bottom of the page
-    # and click the "contact us for help with User's gift" link
-    # AND: Redirected to the contact us form
-    # AND: Fill out the form
-    # AND: Click the "Send" button
-    # THEN: Confirmation page is loaded.
-
-
-@test_case()
-@web
-@expected_failure
-def test_submit_donation_question(web_base_url, selenium):
-    """Test if user could submit question about donation."""
-    # GIVEN: On the OpenStax homepage
-    # WHEN: From the give page
-    # AND: Scroll down to the bottom of the page and
-    # click the "contact us for help with User's gift" link
-    # AND: Redirected to the contact us form
-    # AND: Fill out the form
-    # AND: Click the "Send" button
-    # THEN: Confirmation page is loaded.
 
 
 @test_case()
@@ -549,39 +478,6 @@ def test_view_social_media(web_base_url, selenium):
     # AND: Once completed click confirm
     # THEN: Give page is loaded back. The donate buttons is gone
     # and social media icons (facebook, twitter, linkedin) should be visible.
-
-
-@test_case()
-@web
-@expected_failure
-def test_top_article(web_base_url, selenium):
-    """Test if the top article link functions properly."""
-    # GIVEN: On the OpenStax homepage
-    # WHEN: From the top nav bar find Blog button and click it
-    # AND: On the region below header, click the title of the article
-    # THEN: Full blog article page is loaded.
-
-
-@test_case()
-@web
-@expected_failure
-def test_top_article_read_more(web_base_url, selenium):
-    """Test if the read more link functions properly."""
-    # GIVEN: On the OpenStax homepage
-    # WHEN: From the top nav bar find Blog button and click it
-    # AND: On the region below header, click Read More
-    # THEN: Full blog article page is loaded.
-
-
-@test_case()
-@web
-@expected_failure
-def test_sign_up_from_blog(web_base_url, selenium):
-    """Test if user could use the sign up button from blog."""
-    # GIVEN: On the OpenStax homepage
-    # WHEN: From the top nav bar find Blog button and click it
-    # AND: Find the Sign Up button on the page and click it
-    # THEN:  A new tap of sign up page is loaded
 
 
 @test_case()
